Fix instances of -Wsign-compare for comparisons with constants

+++ iop/iLink/iLinkman/src/iLink.c
@@ -162,7 +162,7 @@ int iLinkTrAlloc(unsigned short int NodeID, unsigned char speed){
 			}
 		}
 
-		result=(i==MAX_CONCURRENT_TRANSACTIONS)?-1002:i;
+		result=(i==MAX_CONCURRENT_TRANSACTIONS)?-1002:(int)i;
 	}
 	else result = -1;
